pyproject.toml
[tool.poetry]
name = "Prototype"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "<2"
tqdm = "^4.66.1"
python-dotenv = "^1.0.1"
langchain-experimental = "^0.0.49"
openai = "^1.10.0"
langchain-openai = "^0.0.5"
unstructured = "^0.12.3"
pdf2image = "^1.17.0"
pdfminer-six = "^20231228"
opencv-python = "^4.9.0.80"
unstructured-inference = "^0.7.23"
pikepdf = "^8.11.2"
pypdf = "^4.0.1"
langchain-community = "^0.0.16"
psycopg = {extras = ["binary", "pool"], version = "^3.1.17"}
pgvector = "^0.2.4"
sentence-transformers = "^2.3.1"
transformers = "^4.37.2"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
